requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
antlr4-python3-runtime==4.9.2
    # via -r requirements.in
astroid==2.5.3
    # via pylint
colorama==0.4.4
    # via pylint
isort==5.8.0
    # via pylint
lazy-object-proxy==1.6.0
    # via astroid
mccabe==0.6.1
    # via pylint
pylint==2.7.4
    # via -r requirements.in
toml==0.10.2
    # via pylint
typed-ast==1.4.3
    # via astroid
wrapt==1.12.1
    # via astroid
